+++ src/games/alienisolation/aliasisolation/runtime/config.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+/*
+ * Compile-time feature toggles for Alias Isolation extras.
+ *
+ * Set either value to 0 from the build or before including the runtime headers
+ * to remove that replacement path from shader identification and pipeline
+ * cloning. TAA, jitter, and the shadow fixes stay independent of these toggles.
+ */
+
+#ifndef ALIENISOLATION_ENABLE_BARREL_DISTORTION_REMOVAL
+#define ALIENISOLATION_ENABLE_BARREL_DISTORTION_REMOVAL 0
+#endif
+
+#ifndef ALIENISOLATION_ENABLE_BLOOM_MERGE_REPLACEMENT
+#define ALIENISOLATION_ENABLE_BLOOM_MERGE_REPLACEMENT 0
+#endif
