+++ src/games/alienisolation/aliasisolation/runtime/shader_ids.hpp
@@ -1,24 +1,37 @@
 #pragma once
 
+/*
+ * DXBC checksum table for the Alien Isolation passes Alias Isolation depends on.
+ *
+ * These are the four 32-bit checksum words stored in the DXBC header, matching
+ * the original ASI's shaderHooks.cpp checks. They are independent of RenoDX's
+ * dumped-file CRC32 names and only identify original game shaders/passes.
+ */
+
 #include <array>
 #include <cstdint>
 #include <cstring>
 #include <span>
 
+#include "./config.hpp"
+
 namespace alienisolation::aliasisolation {
 
 enum class ShaderId : uint32_t {
   Unknown,
+#if ALIENISOLATION_ENABLE_BARREL_DISTORTION_REMOVAL
   MainPostVs,
+#endif
   SmaaVs,
   RgbmEncodeVs,
   RgbmEncodePs,
   DofEncodePs,
   CameraMotionPs,
-  SmaaSpatialPs,
   ShadowLinearizePs,
   ShadowDownsamplePs,
+#if ALIENISOLATION_ENABLE_BLOOM_MERGE_REPLACEMENT
   BloomMergePs,
+#endif
   Count,
 };
 
@@ -30,13 +43,16 @@ inline bool MatchesDxbcChecksum(std::span<const uint8_t> code, const DxbcChecksu
   if (code.size() < 20u) return false;
   if (code[0] != 'D' || code[1] != 'X' || code[2] != 'B' || code[3] != 'C') return false;
 
+  // DXBC bytes 4..19 are the checksum Alias Isolation used for shader matching.
   std::array<uint32_t, 4> actual = {};
   std::memcpy(actual.data(), code.data() + 4u, sizeof(actual));
   return actual == checksum.words;
 }
 
 inline ShaderId IdentifyVertexShader(std::span<const uint8_t> code) {
+#if ALIENISOLATION_ENABLE_BARREL_DISTORTION_REMOVAL
   if (MatchesDxbcChecksum(code, {{0x40af2521, 0x0abe7ef9, 0x8cafc30b, 0xa48433af}})) return ShaderId::MainPostVs;
+#endif
   if (MatchesDxbcChecksum(code, {{0x021e3541, 0xc8808c25, 0x81beb6be, 0x342bfddd}})) return ShaderId::SmaaVs;
   if (MatchesDxbcChecksum(code, {{0xa851671d, 0xbe79cf68, 0x2e6d9376, 0x567ba13c}})) return ShaderId::RgbmEncodeVs;
   return ShaderId::Unknown;
@@ -46,17 +62,20 @@ inline ShaderId IdentifyPixelShader(std::span<const uint8_t> code) {
   if (MatchesDxbcChecksum(code, {{0x4fcfc7f7, 0x5c5e12cf, 0x059e8b33, 0x11f8489b}})) return ShaderId::RgbmEncodePs;
   if (MatchesDxbcChecksum(code, {{0x29ed6504, 0x77d5438c, 0xe9c206c8, 0xb1f27ba2}})) return ShaderId::DofEncodePs;
   if (MatchesDxbcChecksum(code, {{0x1fb3edd4, 0xe984323b, 0x11bcf154, 0x5a029c94}})) return ShaderId::CameraMotionPs;
-  if (MatchesDxbcChecksum(code, {{0x02b5231b, 0x8b3879b8, 0x7db9bc8d, 0xf46a9d78}})) return ShaderId::SmaaSpatialPs;
   if (MatchesDxbcChecksum(code, {{0x3c7b9d08, 0x7b3adf54, 0x3bfc6b9d, 0x1b0ec92e}})) return ShaderId::ShadowLinearizePs;
   if (MatchesDxbcChecksum(code, {{0x8d485646, 0x7d707454, 0x95bf6cb1, 0x09b85460}})) return ShaderId::ShadowDownsamplePs;
+#if ALIENISOLATION_ENABLE_BLOOM_MERGE_REPLACEMENT
   if (MatchesDxbcChecksum(code, {{0xc0fbe484, 0x2d952c03, 0x993de248, 0x56ad9263}})) return ShaderId::BloomMergePs;
+#endif
   return ShaderId::Unknown;
 }
 
 inline const char* ShaderIdName(ShaderId id) {
   switch (id) {
+#if ALIENISOLATION_ENABLE_BARREL_DISTORTION_REMOVAL
     case ShaderId::MainPostVs:
       return "MainPostVs";
+#endif
     case ShaderId::SmaaVs:
       return "SmaaVs";
     case ShaderId::RgbmEncodeVs:
@@ -67,14 +86,14 @@ inline const char* ShaderIdName(ShaderId id) {
       return "DofEncodePs";
     case ShaderId::CameraMotionPs:
       return "CameraMotionPs";
-    case ShaderId::SmaaSpatialPs:
-      return "SmaaSpatialPs";
     case ShaderId::ShadowLinearizePs:
       return "ShadowLinearizePs";
     case ShaderId::ShadowDownsamplePs:
       return "ShadowDownsamplePs";
+#if ALIENISOLATION_ENABLE_BLOOM_MERGE_REPLACEMENT
     case ShaderId::BloomMergePs:
       return "BloomMergePs";
+#endif
     default:
       return "Unknown";
   }
