+++ src/games/alienisolation/aliasisolation/shaders/aliasisolation_shadow_downsample.ps_4_0.hlsl
