+++ src/games/alienisolation/aliasisolation/shaders/aliasisolation_taa.cs_5_0.hlsl
@@ -1,5 +1,3 @@
-#include "../../../shaders/renodx.hlsl"
-
 Texture2D<float4> t3 : register(t3);
 
 Texture2D<float4> t2 : register(t2);
@@ -194,9 +192,9 @@ void main(uint3 vThreadID: SV_DispatchThreadID) {
   // r0.x = dot(float3(0.298999995, 0.587000012, 0.114), r5.rgb);
   // r0.y = dot(float3(0.298999995, 0.587000012, 0.114), r3.rgb);
   // r1.z = dot(float3(0.298999995, 0.587000012, 0.114), r4.rgb);
-  r0.x = renodx::color::yf::from::BT709(r5.rgb);
-  r0.y = renodx::color::yf::from::BT709(r3.rgb);
-  r1.z = renodx::color::yf::from::BT709(r4.rgb);
+  r0.x = dot(float3(0.2126390059f, 0.7151686788f, 0.0721923154f), r5.rgb);
+  r0.y = dot(float3(0.2126390059f, 0.7151686788f, 0.0721923154f), r3.rgb);
+  r1.z = dot(float3(0.2126390059f, 0.7151686788f, 0.0721923154f), r4.rgb);
 
   r1.w = r0.x + -r0.y;
   r0.x = -r1.z + r0.x;
