+++ src/games/alienisolation/tonemap.hlsl
@@ -219,7 +219,6 @@ void GetSceneColorAndTexCoord(
     r2.z = SamplerFrameBuffer_TEX.SampleLevel(SamplerFrameBuffer_SMP_s, r1.zw, 0).z;
   } else {  // chromatic aberration off
     r2.xyz = SamplerFrameBuffer_TEX.SampleLevel(SamplerFrameBuffer_SMP_s, r0.xy, 0).xyz;
-    r2.rgb = ApplyRCAS(r2.rgb, r0.xy, SamplerFrameBuffer_TEX, SamplerFrameBuffer_SMP_s);
   }
 
   scene_color = r2;
@@ -464,7 +463,7 @@ float3 EncodeGamma(float3 linear_color) {
 
 float3 ApplyFilmGrain(float3 input_color, Texture2D<float4> SamplerNoise_TEX,
                       SamplerState SamplerNoise_SMP_s, float4 v1) {
-  float3 grained_color;
+  float3 output_color = input_color;
   if (injectedData.fxFilmGrainType == 0.f) {  // Noise
     float4 r0, r2;
     float3 r1, r3;
@@ -482,27 +481,9 @@ float3 ApplyFilmGrain(float3 input_color, Texture2D<float4> SamplerNoise_TEX,
     r1.xyz = r1.xyz * r2.zzz + r2.xyw;
     r0.xyz = injectedData.fxFilmGrain * r1.xyz + r0.xyz;
 
-    grained_color = r0.rgb;
-  } else {  // Film Grain, applied in linear space
-    float3 linear_color = renodx::color::gamma::DecodeSafe(input_color, 2.2f);
-    if (injectedData.fxFilmGrainType == 1.f) {  // B&W
-      float random_seed = SamplerNoise_TEX.Sample(SamplerNoise_SMP_s, v1.xy).y;
-      grained_color = renodx::effects::ApplyFilmGrain(
-          linear_color,
-          v1.xy,        // Screen-space coordinates
-          random_seed,  // Sample noise tex for random seed
-          injectedData.fxFilmGrain * .04f);
-    } else {  // Colored
-      float3 random_seed = SamplerNoise_TEX.Sample(SamplerNoise_SMP_s, v1.xy).rgb;
-      grained_color = renodx::effects::ApplyFilmGrainColored(
-          linear_color,
-          v1.xy,        // Screen-space coordinates
-          random_seed,  // Sample noise tex for random seed
-          injectedData.fxFilmGrain * .04f);
-    }
-    grained_color = renodx::color::gamma::EncodeSafe(grained_color, 2.2f);
+    output_color = r0.rgb;
   }
-  return grained_color;
+  return output_color;
 }
 
 float3 ApplyBloodOverlay(
